LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource-timing-expected.txt


FAIL No timeline entry for about:blank assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL Setting 'document.domain' does not effect same-origin checks assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'iframe (Populate cache): The initial request populates the cache (if appropriate). assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'iframe (Potentially Cached): Immediately fetch the same URL, exercising the cache hit path (if any). assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'xmlhttprequest (Populate cache): The initial request populates the cache (if appropriate). assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'xmlhttprequest (Potentially Cached): Immediately fetch the same URL, exercising the cache hit path (if any). assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'script (Populate cache): The initial request populates the cache (if appropriate). assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'script (Potentially Cached): Immediately fetch the same URL, exercising the cache hit path (if any). assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'link (Populate cache): The initial request populates the cache (if appropriate). assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'link (Potentially Cached): Immediately fetch the same URL, exercising the cache hit path (if any). assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'iframe: 1 second delay before 'responseStart', another 1 second delay before 'responseEnd'. assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'xmlhttprequest: 1 second delay before 'responseStart', another 1 second delay before 'responseEnd'. assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'script: 1 second delay before 'responseStart', another 1 second delay before 'responseEnd'. assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'link: 1 second delay before 'responseStart', another 1 second delay before 'responseEnd'. assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'iframe (Redirected): 1 second delay before 'redirectEnd', another 1 second delay before 'responseStart'. assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'xmlhttprequest (Redirected): 1 second delay before 'redirectEnd', another 1 second delay before 'responseStart'. assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'script (Redirected): 1 second delay before 'redirectEnd', another 1 second delay before 'responseStart'. assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
FAIL 'link (Redirected): 1 second delay before 'redirectEnd', another 1 second delay before 'responseStart'. assert_less_than_equal: fetchStart <= domainLookupStart expected a number less than or equal to 0 but got 103.4
Show details
